requirements.txt
enum34==1.1.2
PyYAML==3.11
